+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Full Keycloak upstream jira issue can be shown if filtered by Fix version.
 
 Our Keycloak version is working well with PostgreSQL database. For using other SQL databases, text field in database need to be evaluated.
 
-## [Unreleased]
+## [22.0.13-1.17] - 2025-05-28
 
 ### Fixed
 - Do not ask for mfa in first user login (with enabled mfa browser authentication)